factored out config setting, added cumulative reward calculation, added discount factor
